requirements.txt
fastapi
uvicorn
python-dotenv
httpx
openai
sqlalchemy
psycopg[binary] # PostgreSQL + pgvector
pgvector
pydantic
email-validator
google-auth
google-auth-oauthlib